requirements.txt
asgiref==3.2.3
backcall==0.1.0
colorama==0.4.3
coverage==5.1
decorator==4.4.2
Django==3.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-extensions==2.2.9
django-formset-js-improved==0.5.0.2
django-jquery-js==3.1.1
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
Markdown==3.1.1
parso==0.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==3.0.5
Pygments==2.6.1
pytz==2019.3
six==1.15.0
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.2.4
Werkzeug==1.0.1
